requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.1
PyJWT==2.9.0
pytest==8.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.34
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.4
